Move types to the types.ts file
